+++ scilab/modules/fileio/src/c/mgetstr.c
@@ -23,7 +23,7 @@ void C2F(mgetstr) (integer *fd, char **start, integer *n, integer *ierr)
 			*ierr=1;
 			return;
 		}
-		count=fread(*start,sizeof(char),*n,fa);
+		count=(int)fread(*start,sizeof(char),*n,fa);
 		(*start)[*n]='\0';
 		if ( count != *n ) 
 		{
@@ -43,7 +43,7 @@ void C2F(mgetstr1) (integer *fd, char *start, integer *n, integer *ierr)
 	*ierr=0;
 	if (fa)
 	{ 
-		count=fread(start,sizeof(char),*n,fa);
+		count=(int)fread(start,sizeof(char),*n,fa);
 		start[*n]='\0';
 		if ( count != *n ) 
 		{
